# api/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package api

const (
	BearerAuthScopes = "BearerAuth.Scopes"
)

// DefaultErrorResponse defines model for DefaultErrorResponse.
type DefaultErrorResponse struct {
	// Error error description
	Error string `json:"error"`
}

// OrderCreatePostRequest defines model for OrderCreatePostRequest.
type OrderCreatePostRequest struct {
	OrderItems []OrderItemRequest     `json:"orderItems"`
	Shipping   ShippingAddressRequest `json:"shipping"`
}

// OrderCreatePostResponse defines model for OrderCreatePostResponse.
type OrderCreatePostResponse struct {
	// Id id of the order (uuid)
	Id string `json:"id"`
}

// OrderDetailGetResponse defines model for OrderDetailGetResponse.
type OrderDetailGetResponse struct {
	// Id id of the order
	Id         string              `json:"id"`
	OrderItems []OrderItemResponse `json:"orderItems"`

	// OrderTime order creation time (in unix millis)
	OrderTime int64                  `json:"orderTime"`
	Shipping  ShippingAddressRequest `json:"shipping"`

	// Status order status
	Status string `json:"status"`
}

// OrderItemRequest defines model for OrderItemRequest.
type OrderItemRequest struct {
	// Amount order amount
	Amount int `json:"amount" validate:"required"`

	// ProductId id of the product
	ProductId string `json:"productId" validate:"required,uuid"`
}

// OrderItemResponse defines model for OrderItemResponse.
type OrderItemResponse struct {
	// Id id of the order item
	Id string `json:"id"`

	// ProductImage product image in full accessible url
	ProductImage string `json:"productImage"`

	// ProductName product name
	ProductName string `json:"productName"`

	// ProductPrice product price
	ProductPrice int64 `json:"productPrice"`

	// Qty order quantity for this product
	Qty int `json:"qty"`
}

// OrderListGetDetail defines model for OrderListGetDetail.
type OrderListGetDetail struct {
	// Id id of the order
	Id string `json:"id"`

	// Status latest order status
	Status string `json:"status"`

	// TotalAmount total order amount
	TotalAmount int64 `json:"totalAmount"`
}

// OrderListGetResponse defines model for OrderListGetResponse.
type OrderListGetResponse struct {
	Data       []OrderListGetDetail `json:"data"`
	Pagination PaginationSchema     `json:"pagination"`
}

// OrderStatusPutRequest defines model for OrderStatusPutRequest.
type OrderStatusPutRequest struct {
	// Status order status to be updated
	Status string `json:"status" validate:"required,oneof=confirmed"`
}

// PaginationSchema defines model for PaginationSchema.
type PaginationSchema struct {
	// Page current active page
	Page int `json:"page"`

	// PageSize number of active data in the page
	PageSize int `json:"pageSize"`

	// TotalData total data available to be outputted. meaning that max page is 1000 (totalData / pageSize)
	TotalData int64 `json:"totalData"`
}

// ProductDetailGetResponse defines model for ProductDetailGetResponse.
type ProductDetailGetResponse struct {
	// Description Product description
	Description string `json:"description"`

	// Dimension Product dimension
	Dimension string `json:"dimension"`

	// DiscountedPrice Product discounted price
	DiscountedPrice int `json:"discountedPrice"`

	// Id id of product
	Id string `json:"id"`

	// Image Product image
	Image string `json:"image"`

	// Medium Art medium
	Medium string `json:"medium"`

	// Name Product name
	Name string `json:"name"`

	// OriginalPrice Product original price
	OriginalPrice int `json:"originalPrice"`

	// Rating Product rating
	Rating int `json:"rating"`
}

// ProductListGetResponse defines model for ProductListGetResponse.
type ProductListGetResponse = []ProductListGetResponseItem

// ProductListGetResponseItem defines model for ProductListGetResponseItem.
type ProductListGetResponseItem struct {
	// DiscountedPrice Product discounted price
	DiscountedPrice int `json:"discountedPrice"`

	// Id id of product
	Id string `json:"id"`

	// Image Product image
	Image string `json:"image"`

	// Name Product name
	Name string `json:"name"`

	// OriginalPrice Product original price
	OriginalPrice int `json:"originalPrice"`

	// Rating Product rating
	Rating int `json:"rating"`
}

// ShippingAddressRequest defines model for ShippingAddressRequest.
type ShippingAddressRequest struct {
	// FullAddress receiver full address
	FullAddress string `json:"fullAddress" validate:"required"`

	// FullName contactable receiver full name
	FullName string `json:"fullName" validate:"required"`

	// Notes additional shipping notes
	Notes *string `json:"notes,omitempty"`

	// PhoneNumber contactable receiver phone number
	PhoneNumber string `json:"phoneNumber" validate:"required"`
}

// OptionalPageParam defines model for OptionalPageParam.
type OptionalPageParam = int

// OrderIdPathParams defines model for OrderIdPathParams.
type OrderIdPathParams = string

// ProductIdPathParams defines model for ProductIdPathParams.
type ProductIdPathParams = string

// ProductListGetParams defines parameters for ProductListGet.
type ProductListGetParams struct {
	// Page paginate result
	Page *OptionalPageParam `form:"page,omitempty" json:"page,omitempty" validate:"omitempty,gt=0"`
}

// OrderListGetJSONRequestBody defines body for OrderListGet for application/json ContentType.
type OrderListGetJSONRequestBody = OrderCreatePostRequest

// OrderCreatePostJSONRequestBody defines body for OrderCreatePost for application/json ContentType.
type OrderCreatePostJSONRequestBody = OrderCreatePostRequest

// OrderStatusPutJSONRequestBody defines body for OrderStatusPut for application/json ContentType.
type OrderStatusPutJSONRequestBody = OrderStatusPutRequest