# api/api.yaml
openapi: 3.0.0
info:
  title: Besart Gallery API
  description: API to perform BESArt gallery functionality
  version: 1.0.0
tags:
  - name: auth
    description: Endpoints for authentication
  - name: products
    description: Endpoints for products
  - name: order
    description: Endpoints for order management
paths:
  /products:
    get:
      tags:
        - products
      summary: Get product listing
      description: Get product listing
      operationId: ProductListGet
      parameters:
        - $ref: '#/components/parameters/OptionalPageParam'
      responses:
        '200':
          description: Success - Return product listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListGetResponse'
        '500':
          description: Error - Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultErrorResponse'
  /products/{product_id}:
    get:
      tags:
        - products
      summary: Get product detail
      description: Get product detail
      operationId: ProductDetailGet
      parameters:
        - $ref: '#/components/parameters/ProductIdPathParams'
      responses:
        '200':
          description: Success - Return product detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetailGetResponse'
        '404':
          description: Error - Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultErrorResponse'
  /orders:
    post:
      tags:
        - order
      summary: Create new order
      description: Create new order (open)
      operationId: OrderCreatePost
      security:
        - BearerAuth: ['create-order']
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreatePostRequest'
      responses:
        '200':
          description: Success - Return order id and number (promised)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderCreatePostResponse'
        '404':
          description: Error - Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultErrorResponse'
    get:
      tags:
        - order
      summary: Get order list
      description: Listing of created order by authenticated user
      operationId: OrderListGet
      security:
        - BearerAuth: ['list-order']
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreatePostRequest'
      responses:
        '200':
          description: Success - Return order id and number (promised)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListGetResponse'
        '500':
          description: Error - Unknown Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultErrorResponse'
  /orders/{order_id}:
    get:
      tags:
        - order
      summary: Get order detail
      description: Get order detail
      operationId: OrderDetailGet
      parameters:
        - $ref: '#/components/parameters/OrderIdPathParams'
      responses:
        '200':
          description: Success - Return product detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailGetResponse'
        '404':
          description: Error - Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultErrorResponse'      
  /orders/{order_id}/status:
    put:
      tags:
        - order
      summary: Update order status
      description: Update order status
      operationId: OrderStatusPut
      security:
        - BearerAuth: ['update-order-status']
      parameters:
        - $ref: '#/components/parameters/OrderIdPathParams'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderStatusPutRequest'
      responses:
        '200':
          description: Success - Return order id and number (promised)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderCreatePostResponse'
        '404':
          description: Error - Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        Authorization using header parameter with parameter name `Authorization`
        See: https://swagger.io/docs/specification/authentication/bearer-authentication/
      bearerFormat: Bearer <token>
  parameters:
    OptionalPageParam:
      name: page
      in: query
      required: false
      description: paginate result
      schema:
        type: integer
      x-oapi-codegen-extra-tags:
        validate: omitempty,gt=0
    ProductIdPathParams:
      name: product_id
      in: path
      required: true
      description: id of the product
      schema:
        type: string
      x-oapi-codegen-extra-tags:
        validate: required,uuid
    OrderIdPathParams:
      name: order_id
      in: path
      required: true
      description: id of the order
      schema:
        type: string
      x-oapi-codegen-extra-tags:
        validate: required,uuid
  schemas:
    DefaultErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: error description
          example: failed to validate nama field
    ProductListGetResponse:
      type: array
      items:
        $ref: '#/components/schemas/ProductListGetResponseItem'
    ProductListGetResponseItem:
      type: object
      required:
        - id
        - name
        - image
        - rating
        - discountedPrice
        - originalPrice
      properties:
        id:
          type: string
          description: id of product
          example: 32bf67e5-4971-47ce-985c-44b6b3860cdb
        name:
          type: string
          description: Product name
          example: Sunset Over Still Waters
        image:
          type: string
          description: Product image
          example: https://template.getbazaar.io/_next/image?url=%2Fassets%2Fimages%2Ffurniture-products%2Ffurniture-8.png&w=828&q=75
        rating:
          type: integer
          description: Product rating
          example: 4
          format: float64
        discountedPrice:
          type: integer
          description: Product discounted price
          example: 10000000
        originalPrice:
          type: integer
          description: Product original price
          example: 15000000
    ProductDetailGetResponse:
      type: object
      required:
        - id
        - name
        - image
        - rating
        - discountedPrice
        - originalPrice
        - description
        - dimension
        - medium
      properties:
        id:
          type: string
          description: id of product
          example: 32bf67e5-4971-47ce-985c-44b6b3860cdb
        name:
          type: string
          description: Product name
          example: Sunset Over Still Waters
        image:
          type: string
          description: Product image
          example: https://template.getbazaar.io/_next/image?url=%2Fassets%2Fimages%2Ffurniture-products%2Ffurniture-8.png&w=828&q=75
        rating:
          type: integer
          description: Product rating
          example: 4
          format: float64
        discountedPrice:
          type: integer
          description: Product discounted price
          example: 10000000
        originalPrice:
          type: integer
          description: Product original price
          example: 15000000
        description:
          type: string
          description: Product description
          example: Lorem ipsum dolor sit amet\nconsectetur adipiscing elit\nsed do eiusmod tempor incididunt ut labore et dolore magna aliqua
        dimension:
          type: string
          description: Product dimension
          example: 120 x 80 cm
        medium:
          type: string
          description: Art medium
          example: Oil Painting on Canvas
    OrderCreatePostRequest:
      type: object
      required:
        - shipping
        - orderItems
      properties:
        shipping:
          $ref: '#/components/schemas/ShippingAddressRequest'
        orderItems:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemRequest'
    ShippingAddressRequest:
      type: object
      required:
        - fullName
        - phoneNumber
        - fullAddress
      properties:
        fullName:
          type: string
          description: contactable receiver full name
          example: John Doe
          x-oapi-codegen-extra-tags:
            validate: required
        phoneNumber:
          type: string
          description: contactable receiver phone number
          example: "+6281123456789"
          x-oapi-codegen-extra-tags:
            validate: required
        fullAddress:
          type: string
          description: receiver full address
          example: 123 Giggle Ave, Richmond, VA 22548
          x-oapi-codegen-extra-tags:
            validate: required
        notes:
          type: string
          description: additional shipping notes
          example: please just put on terrace
    OrderItemRequest:
      type: object
      required:
        - productId
        - amount
      properties:
        productId:
          type: string
          description: id of the product
          example: bf6b22f4-ed76-4872-acbd-43f3910c796e
          x-oapi-codegen-extra-tags:
            validate: required,uuid
        amount:
          type: integer
          description: order amount
          example: 1
          x-oapi-codegen-extra-tags:
            validate: required
    OrderCreatePostResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          description: id of the order (uuid)
          example: bf6b22f4-ed76-4872-acbd-43f3910c796e
    OrderStatusPutRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          description: order status to be updated
          example: confirmed
          x-oapi-codegen-extra-tags:
            validate: required,oneof=confirmed
    OrderListGetResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/OrderListGetDetail'
        pagination:
          $ref: '#/components/schemas/PaginationSchema'
    OrderListGetDetail:
      type: object
      required:
        - id
        - totalAmount
        - status
      properties:
        id:
          type: string
          description: id of the order
          example: 1a2af919-ac78-4e1b-93ac-4b46482422aa
        totalAmount:
          type: integer
          format: int64
          description: total order amount
          example: 10000000
        status:
          type: string
          description: latest order status
          example: open
    PaginationSchema:
      type: object
      required:
        - page
        - pageSize
        - totalData
      properties:
        page:
          type: integer
          description: current active page
          example: 1
        pageSize:
          type: integer
          description: number of active data in the page
          example: 10
        totalData:
          type: integer
          format: int64
          description: total data available to be outputted. meaning that max page is 1000 (totalData / pageSize)
          example: 10000
    OrderDetailGetResponse:
      type: object
      required:
        - id
        - orderTime
        - status
        - shipping
        - orderItems
      properties:
        id:
          type: string
          description: id of the order
          example: 1a2af919-ac78-4e1b-93ac-4b46482422aa
        orderTime:
          type: integer
          format: int64
          description: order creation time (in unix millis)
          example: 1749703960000
        status:
          type: string
          description: order status
          example: open
        shipping:
          $ref: '#/components/schemas/ShippingAddressRequest'
        orderItems:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
    OrderItemResponse:
      type: object
      required:
        - id
        - productImage
        - productName
        - qty
        - productPrice
      properties:
        id: 
          type: string
          description: id of the order item
          example: 887e8ac8-98d5-4823-88ec-a59fa804b7af
        productImage:
          type: string
          description: product image in full accessible url
          example: https://besarts.biz.id/assets/image.jpg
        productName:
          type: string
          description: product name
          example: Some Product Name
        qty:
          type: integer
          description: order quantity for this product
          example: 1
        productPrice:
          type: integer
          format: int64
          description: product price
          example: 10000000